% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_EDF.R
\name{calculate_EDF}
\alias{calculate_EDF}
\title{Calculate effective degrees of freedom code from
Thorson JT. Measuring complexity for hierarchical models using effective
degrees of freedom. Ecology. 2024 Jun 11:e4327. doi: 10.1002/ecy.4327.
Epub ahead of print. PMID: 38859712.}
\usage{
calculate_EDF(
  obj,
  opt,
  nonvariance_fixed_effects = "",
  prediction_name = "pred_logRS",
  data_name = "obs_logRS",
  delta = 0.1,
  show_progress = FALSE,
  refit = c("random", "full")[1],
  what = c("EDF", "grad_i", "all")[1]
)
}
\arguments{
\item{obj}{tmb_obj output from samEst TMB function}

\item{opt}{model output from samEst TMB function}

\item{nonvariance_fixed_effects}{names of the parameters that are the non varinace
fixed effects in models with random effects}

\item{prediction_name}{predicted variable name}

\item{data_name}{observed y value}

\item{delta}{see thorson paper}

\item{refit}{refit for all data points}

\item{what}{what to output, "EDF": just EDF, "grad_i": individual
gradients for each data point, or "all"}
}
\value{
Sgen likelihood
}
\description{
Calculate effective degrees of freedom code from
Thorson JT. Measuring complexity for hierarchical models using effective
degrees of freedom. Ecology. 2024 Jun 11:e4327. doi: 10.1002/ecy.4327.
Epub ahead of print. PMID: 38859712.
}
