% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_run_functions.R
\name{compile_code}
\alias{compile_code}
\title{compile stan models}
\usage{
compile_code(
  type = c("static", "rw", "hmm"),
  ac = FALSE,
  par = c("n", "a", "b", "both"),
  lambertW = FALSE
)
}
\arguments{
\item{lambertW}{Logical, indicating if lambertW functions should be used in stan code, requires git installation from stan}

\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{AC}{Logical. Are residuals autocorrelated? Default is FALSE}

\item{control}{output of stancontrol}

\item{warmup}{To be passed to rstan::sampling. A positive integer specifying the number of warmup (aka burnin) iterations per
chain. The default is 200.}

\item{chains}{To be passed to rstan::sampling. A positive integer specifying the number of Markov chains. The default is 6.}

\item{iter}{To be passed to rstan::sampling. A positive integer specifying the number of iterations for each chain
(including warmup). The default is 1000.}

\item{...}{Anything else that would be passed to rstan::sampling}
}
\value{
a list containing the
\itemize{
\item alpha - median estimates for the alpha parameter vector
\item beta - median estimates for the beta parameter
\item sigobs - median estimates for the observation error sigma
\item stanfit - a stanfit model object
\item mcmcsummary - summary over kept samples
\item c_mcmcsummary - chain specific summary
\item list of samples
}
}
\description{
compile stan models
}
\examples{
data(harck)
rickerstan(data=harck)

}
