% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LamW_RTMB.r
\name{dLambertW0_internal}
\alias{dLambertW0_internal}
\title{Derivatives of LamW}
\usage{
dLambertW0_internal(x, y, dy)
}
\arguments{
\item{x}{}

\item{y}{}

\item{dy}{}
}
\value{
LambertW derivative
}
\description{
Derivatives of LamW
}
\details{
This is what you would import. Only defined for a scalar. Need to update code if you want to be able to do more than 1 x at a time.
}
