% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{dummystan}
\alias{dummystan}
\title{stanRoxygen commands}
\usage{
dummystan()
}
\description{
This is a dummy function to hold the useDynLib roxygen tag.
This tag will populate the namespace with compiled c++ functions upon package install.
}
