% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{find_linit}
\alias{find_linit}
\title{Find lalpha intial value for logistic transformed TMB HMM model parameters}
\usage{
find_linit(U, L, initpar)
}
\arguments{
\item{U}{a number. upper limit for parameter}

\item{L}{a number. lower limit for parameter}

\item{initpar}{Inital value of parameter to be transformed}
}
\value{
logistic transformed parameter guess
}
\description{
Find lalpha intial value for logistic transformed TMB HMM model parameters
}
