% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LamW_RTMB.r
\name{FritschIter}
\alias{FritschIter}
\title{Fritsch iteration}
\usage{
FritschIter(x, w)
}
\arguments{
\item{x}{.}

\item{w}{.}
}
\value{
lambertW solution
}
\description{
Fritsch iteration
}
\details{
This function was adapted from those in the \href{https://github.com/cran/lamW/blob/master/src/lambertW.cpp}{lamW package}
}
\references{
Corless, R. M.; Gonnet, G. H.; Hare, D. E.; Jeffrey, D. J. & Knuth, D. E.
"On the Lambert W function", Advances in Computational Mathematics,
Springer, 1996, 5, 329-359

Fritsch, F. N.; Shafer, R. E. & Crowley, W. P.
"Solution of the transcendental equation (we^w = x)",
Communications of the ACM, Association for Computing Machinery (ACM),
1973, 16, 123-124
}
