% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harck.R
\docType{data}
\name{harck}
\alias{harck}
\title{Harrison chinook data}
\format{
An object of class \code{data.frame} with 33 rows and 4 columns.
}
\source{
Chuck Parken, DFO
}
\usage{
data(harck)
}
\description{
This dataset includes Total Spawners, S, AEQ adjusted recruitment, R, and log recruits por spawner, logRS.
}
\keyword{datasets}
