% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LamW_RTMB.r
\name{LambertW0}
\alias{LambertW0}
\title{LambertW function to be used inside RTMB models}
\usage{
LambertW0(x)
}
\value{
penalized (if priors_flag= TRUE) negative log likelihood
}
\description{
LambertW function to be used inside RTMB models
}
\details{
code adapted from \href{https://github.com/kaskr/RTMB/blob/d4f7db7f6c3073d7f85930ed6d6d4dd264b05612/RTMB/R/adjoint.R}{adjoint example}
}
