% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LamW_RTMB.r
\name{LambertW0_internal}
\alias{LambertW0_internal}
\title{Lambert W defined for the entire W0 interval}
\usage{
LambertW0_internal(x)
}
\arguments{
\item{x}{a numeric value to apply the lambert W function to.}
}
\value{
lambertW solution
}
\description{
Lambert W defined for the entire W0 interval
}
\details{
this version is also defined for small negative values. Adapted from the
\href{https://github.com/cran/lamW/blob/master/src/lambertW.cpp}{lamW package}. See copyright info in the package
}
