% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{log10_ceiling}
\alias{log10_ceiling}
\title{Find order of magnitude}
\usage{
log10_ceiling(x)
}
\arguments{
\item{x}{a number.}
}
\value{
order of magnitude
}
\description{
Find order of magnitude
}
