% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{log_prior_params}
\alias{log_prior_params}
\title{function to generate lognormal prior parameters based on mean and standard deviation in normal
space. Based on Quan and Zhang (2003) Estimate of standard deviation for a log-transformed
variable using arithmetic means and standard deviations. Statist. Med. 22:2723–2736}
\usage{
log_prior_params(Smax_mean, Smax_sd)
}
\arguments{
\item{Smax_mean}{}

\item{Smax_sd}{}
}
\value{
list of mean and sd for lognormal prior on beta
}
\description{
function to generate lognormal prior parameters based on mean and standard deviation in normal
space. Based on Quan and Zhang (2003) Estimate of standard deviation for a log-transformed
variable using arithmetic means and standard deviations. Statist. Med. 22:2723–2736
}
