% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{minus_one_to_one}
\alias{minus_one_to_one}
\title{function to trasnform parameter between minus one an one}
\usage{
minus_one_to_one(x)
}
\arguments{
\item{x}{a value}
}
\value{
inverse logit transformed variable
}
\description{
function to trasnform parameter between minus one an one
}
