% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{prior_check}
\alias{prior_check}
\title{Prior predictive check}
\usage{
prior_check(
  data,
  type = c("static", "rw"),
  AC = FALSE,
  par = c("a", "b", "both"),
  control = stancontrol()
)
}
\arguments{
\item{fit}{A model fit in rstan from e.g. ricker_stan, ricker_rw_stan, ricker_hmm_stan}
}
\description{
Any arguments to pass to \link[rstan:stanmodel-method-sampling]{rstan::sampling}.
}
