% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTMB_nll_functions.R
\name{ricker_ac_RTMB_fn}
\alias{ricker_ac_RTMB_fn}
\title{Simple Ricker AR1 model function to be called with RTMB}
\usage{
ricker_ac_RTMB_fn(param, dat)
}
\arguments{
\item{par}{A list or data frame containing parameter names and initial guesses}
}
\value{
penalized (if priors_flag= TRUE) negative log likelihood
}
\description{
Simple Ricker AR1 model function to be called with RTMB
}
\details{
this function will
}
