% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_run_functions.R
\name{ricker_hmm_stan}
\alias{ricker_hmm_stan}
\title{random walks Ricker model estimated with stan}
\usage{
ricker_hmm_stan(
  data,
  par = c("a", "b", "both"),
  k_regime = 2,
  control = stancontrol(),
  warmup = 300,
  chains = 6,
  iter = 1000,
  mod = NULL,
  lambertW = FALSE,
  ...
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{par}{Which parameter should vary? Either productivity (intercept, a), capacity (slope, b) or both parameters}

\item{control}{output of stancontrol}

\item{warmup}{To be passed to rstan::sampling. A positive integer specifying the number of warmup (aka burnin) iterations per
chain. The default is 200.}

\item{chains}{To be passed to rstan::sampling. A positive integer specifying the number of Markov chains. The default is 6.}

\item{iter}{To be passed to rstan::sampling. A positive integer specifying the number of iterations for each chain
(including warmup). The default is 1000.}

\item{lambertW}{Logical, indicating if lambertW functions should be used in stan code, requires git installation from stan}

\item{...}{Anything else that would be passed to rstan::sampling}

\item{sm_ext}{Default is null, external stan hmm model. Implemented to speed up simulation evaluation, use with caution.}
}
\value{
a list containing the
\itemize{
\item alpha - median estimates for the alpha parameter vector
\item beta - median estimates for the beta parameter
\item sigobs - median estimates for the observation error sigma
\item stanfit - a stanfit model object
\item mcmcsummary - summary over kept samples
\item c_mcmcsummary - chain specific summary
\item list of samples
}
}
\description{
random walks Ricker model estimated with stan
}
\examples{
data(harck)
ricker_hmm_stan(data=harck)

}
