% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.R
\name{ricker_hmm_TMB}
\alias{ricker_hmm_TMB}
\title{Ricker hidden markov model with regime shiftts for alpha and beta.}
\usage{
ricker_hmm_TMB(
  data,
  tv.par = c("a", "b", "both"),
  k_regime = 2,
  alpha_limits = c(0, 20),
  beta_upper = 0.1,
  silent = FALSE,
  control = TMBcontrol(),
  ini_param = NULL,
  tmb_map = list(),
  priors_flag = 1,
  stan_flag = 0,
  sig_p_sd = 1
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{tv.par}{Which parameters should vary? Either productivity (intercept, a), capacity (slope, b) or both parameters}

\item{k_regime}{Number of regimes to be considered}

\item{alpha_limits}{vector containing two values: upper and lower limit for alpha parameters. default is c(0,20)}

\item{beta_upper}{upper limit for beta parameter. default is 1}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from TMBcontrol() function, to be passed to nlminb()}

\item{ini_param}{Optional. A list with initial parameter guesses. The list should contain: lalpha (vector of length k_regime),
lbeta (vector of length k_regime), logsigma, pi1_tran (vector of length k_regime-1),
and qij_tran (matrix with nrow=k_regime, and ncol=k_regime-1). Keep in mind that the main
parameters (alpha and beta) are transformed with a logistic function to account for the custom upper and lower values.}

\item{priors_flag}{Integer, 1 priors are included in estimation model, 0 priors are not included.
See details for priors documentation. See details for priors documentation.}

\item{stan_flag}{Integer, flag indicating wether or not TMB code will be used with TMBstan - Jacobian
adjustment implemented. Default is 0, jacobian adjustment not included.}

\item{sig_p_sd}{sd for half normal prior on sigma parameter. default is 1.}
}
\value{
a list containing several model outputs:
\itemize{
\item alpha - MLE estimates for the alpha parameter vector
\item beta - MLE estimates for the beta parameter
\item sig  - MLE estimates for the observation error sigma
\item pi  - MLE estimates for the initial state probabilities vector of length k_regime
\item A  - MLE estimates for the transition probabilities, matrix (k_regime x k_regime)
\item AICc - AICc values, given by 2\emph{nll + 2}npar +(2\emph{npar}(npar+1)/(nrow(data)-npar-1)), excluding prior components
\item BIC - BIC values, excluding prior components
\item model - opt object, generated by the \code{TMB::MakeADFun()}
\item tmb_data - data provided to TMB,
\item tmb_params = parameter intial guesses,
\item tmb_map    = mapping indicating which parameters should be fixed or estimated,
\item tmb_random = tmb_random,
\item tmb_obj    = tmb_obj,
\item gradients  = conv$final_grads,
\item bad_eig    = conv$bad_eig,
\item call       = match.call(expand.dots = TRUE),
\item sd_report  = sd_report),
\item class      = "Ricker_tva"
}

@references{
\insertRef{tangIdentificationRecruitmentRegime2021}{samEst}
}
}
\description{
Ricker hidden markov model with regime shiftts for alpha and beta.
}
\details{
This model was published in Tang et al. 2021 Identification of recruitment regime shifts with a hidden Markov stock-recruitment model.
The code for this model was a contribution by Xiaozhuo Tang.

Priors: Weakly informative priors are included for the main parameterst of the model:
alpha ~ gamma(3,1)
logbeta ~ N(-12,3)
sigobs ~ gamma(2,1/3)
qi ~ gamma(2,1/3)
sigb ~ gamma(2,1/3)
}
\examples{
data(harck)
ricker_HMM_TMB(data=harck)
}
