% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.R
\name{ricker_HMM_TMB_a}
\alias{ricker_HMM_TMB_a}
\title{Ricker hidden markov model with regime shiftts for alpha only.}
\usage{
ricker_HMM_TMB_a(
  data,
  k_regime = 2,
  alpha_limits = c(0, 20),
  beta_upper = 0.1,
  sigma_upper = 2,
  silent = FALSE,
  control = TMBcontrol(),
  ini_param = NULL,
  tmb_map = list()
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{k_regime}{Number of regimes to be considered}

\item{alpha_limits}{vector containing two values: upper and lower limit for alpha parameters. default is c(0,20)}

\item{beta_upper}{upper limit for beta parameter. default is 1}

\item{sigma_upper}{upper limit for beta parameter. default is 2}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from TMBcontrol() function, to be passed to nlminb()}

\item{ini_param}{Optional. A list with initial parameter guesses. The list should contain: lalpha (vector of length k_regime),
lbeta (single), lsigma (vector of length k_regime), pi1_tran (vector of length k_regime-1),
and qij_tran (matrix with nrow=k_regime, and ncol=k_regime-1). Keep in mind that the main
parameters (alpha, beta and sigma) are transformed with a logistic function to account for the custom upper and lower values.}
}
\value{
a list containing several model outputs:
\itemize{
\item alpha - MLE estimates for the alpha parameter vector
\item beta - MLE estimate for the beta parameter
\item sig  - MLE estimates for the observation error sigma
\item siga  - MLE estimates for the process error (variation in alpha) sigma
\item model - opt object, generated by the \code{TMB::MakeADFun()}
\item tmb_data - data provided to TMB,
\item tmb_params = parameter intial guesses,
\item tmb_map    = mapping indicating which parameters should be fixed or estimated,
\item tmb_random = tmb_random,
\item tmb_obj    = tmb_obj,
\item gradients  = conv$final_grads,
\item bad_eig    = conv$bad_eig,
\item call       = match.call(expand.dots = TRUE),
\item sd_report  = sd_report),
\item class      = "Ricker_tva"
}

@references{
\insertRef{tangIdentificationRecruitmentRegime2021}{samEst}
}
}
\description{
Ricker hidden markov model with regime shiftts for alpha only.
}
\details{
This model was published in Tang et al. 2021 Identification of recruitment regime shifts with a hidden Markov stock-recruitment model.
The code for this model was a contribution by Xiaozhuo Tang.
}
\examples{
data(harck)
ricker_HMM_TMB(data=harck)
}
