% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.R
\name{ricker_kf_TMB}
\alias{ricker_kf_TMB}
\title{Ricker model with time-varying logalpha estimated with Kalman Filter}
\usage{
ricker_kf_TMB(
  data,
  silent = FALSE,
  control = TMBcontrol(),
  tmb_map = list(),
  fullLL = FALSE
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from TMBcontrol() function, to be passed to nlminb()}

\item{tmb_map}{optional, mapping list indicating if parameters should be estimated of fixed.
Default is all parameters are estimated}

\item{fullLL}{Should full likelihood function be used, should be used if comparing wth AIC.Default is FALSE}
}
\value{
a list containing several model outputs:
\itemize{
\item alpha - KF estimates for the alpha parameter vector
\item beta - KF  estimates for the beta parameter
\item sigobs - MLE estimates for the observation error sigma
\item model - opt object, generated by the \code{stats::nlminb()}
\item tmb_data - data provided to TMB,
\item tmb_params - parameter intial guesses,
\item tmb_map - mapping indicating which parameters should be fixed or estimated,
\item tmb_obj - obj object, generated by the \code{TMB::MakeADFun()}
\item gradients - final convergence gradient
\item bad_eig - eigen value
\item call - original function call
\item sd_report - MLE estimates and sdt Error estimates for main parameter estimates
\item class - name of cpp model
}
}
\description{
Ricker model with time-varying logalpha estimated with Kalman Filter
}
\examples{
data(harck)
rickerTMB(data=harck)

}
