% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_run_functions.R
\name{ricker_rw_stan}
\alias{ricker_rw_stan}
\title{random walks Ricker model estimated with stan}
\usage{
ricker_rw_stan(
  data,
  tv.par = c("a", "b", "both"),
  smax_priors = NULL,
  control = stancontrol(adapt_delta = 0.99),
  mod = NULL,
  warmup = 300,
  chains = 6,
  iter = 1000,
  ...
)
}
\arguments{
\item{data}{A list or data frame containing complete vectors for: brood year (by), Spawners (s) and log(Recruits/Spawners) (logRS) time series. Use sr_format for the correct column names.}

\item{tv.par}{Which parameter should vary? Either productivity (intercept, a), capacity (slope, b) or both parameters (both)}

\item{smax_priors}{User-specified smax priors, sampled from a normal distribution, should be formatted as: c(mean, sd)}

\item{control}{output of stancontrol}

\item{warmup}{To be passed to rstan::sampling. A positive integer specifying the number of warmup (aka burnin) iterations per
chain. The default is 200.}

\item{chains}{To be passed to rstan::sampling. A positive integer specifying the number of Markov chains. The default is 6.}

\item{iter}{To be passed to rstan::sampling. A positive integer specifying the number of iterations for each chain
(including warmup). The default is 1000.}

\item{...}{Anything else that would be passed to rstan::sampling}
}
\value{
a list containing the
\itemize{
\item data - input dataseries
\item logalpha - median estimates for the log(alpha) parameter
\item beta - median estimates for the beta parameter
\item Smax - median estimates for the Smax parameter
\item Smsy - median estimates for the Smsy parameter
\item Umsy - median estimates for the Umsy parameter
\item sigma - median estimates for the sigma parameter
\item sigma_a - median estimates for the sigma_a parameter (variance attributed to random walk in log(alpha))
\item sigma_b - median estimates for the sigma_b parameter (variance attributed to random walk in Smax)
\item residuals - median estimates for each epsilon (residual) parameter
\item summary - summary of stan model fit
\item full_posterior - all posterior samples of the above parameters
\item stanfit - the stanfit model object
}
}
\description{
random walks Ricker model estimated with stan
}
\examples{
data(harck)
ricker_rw_stan(data=harck)

}
