% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.R
\name{ricker_rwb_TMB}
\alias{ricker_rwb_TMB}
\title{Ricker model with random walk in log(beta) parameter with TMB}
\usage{
ricker_rwb_TMB(
  data,
  silent = FALSE,
  control = TMBcontrol(),
  ini_param = NULL,
  tmb_map = list()
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from TMBcontrol() function, to be passed to nlminb()}

\item{ini_param}{Optional. A list with initial parameter guesses. The list should contain: logbetao (a number),
alpha (a number), logsigobs (a number), logsigb (a number), and logbeta ( a vector with the same length as the data).}

\item{tmb_map}{optional, mapping list indicating if parameters should be estimated of fixed.
Default is all parameters are estimated}
}
\value{
a list containing several model outputs:
\itemize{
\item alpha - MLE estimates for the alpha parameter
\item beta - MLE estimates for the beta parameter vector
\item sig - MLE estimates for the observation error sigma
\item sigb - MLE estimates for the process error (variation in log(beta)) sigma
\item model - opt object, generated by the \code{stats::nlminb()}
\item tmb_data - data provided to TMB,
\item tmb_params - parameter intial guesses,
\item tmb_map - mapping indicating which parameters should be fixed or estimated,
\item tmb_obj - obj object, generated by the \code{TMB::MakeADFun()}
\item gradients - final convergence gradient
\item bad_eig - eigen value
\item call - original function call
\item sd_report - MLE estimates and sdt Error estimates for main parameter estimates
\item class - name of cpp model
}
}
\description{
Ricker model with random walk in log(beta) parameter with TMB
}
\examples{
data(harck)
ricker_rwb_TMB(data=harck)

}
