% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_run_functions.R
\name{ricker_stan}
\alias{ricker_stan}
\title{Simple Ricker model estimated with stan}
\usage{
ricker_stan(
  data,
  ac = FALSE,
  Smax_mean = NULL,
  Smax_sd = NULL,
  mod = NULL,
  control = stancontrol(adapt_delta = 0.99),
  warmup = 300,
  chains = 6,
  iter = 1000,
  ...
)
}
\arguments{
\item{data}{A list or data frame containing complete vectors for: brood year (by), Spawners (s) and log(Recruits/Spawners) (logRS) time series.}

\item{ac}{Logical. Are residuals autocorrelated? Default is FALSE}

\item{Smax_mean}{User-specified smax priors for mean, sampled from a normal distribution. Defaults to mean = half maximum observed spawners.}

\item{Smax_sd}{User-specified smax priors for std. deviation, sampled from a normal distribution. defaults to sd = 2x max. observed spawners (from time series)}

\item{control}{output of stancontrol}

\item{warmup}{To be passed to rstan::sampling. A positive integer specifying the number of warmup (aka burnin) iterations per
chain. The default is 200.}

\item{chains}{To be passed to rstan::sampling. A positive integer specifying the number of Markov chains. The default is 6.}

\item{iter}{To be passed to rstan::sampling. A positive integer specifying the number of iterations for each chain
(including warmup). The default is 1000.}

\item{...}{Anything else that would be passed to rstan::sampling}
}
\value{
a list containing the
\itemize{
\item data - input dataseries
\item logalpha - median estimates for the log(alpha) parameter
\item beta - median estimates for the beta parameter
\item Smax - median estimates for the Smax parameter
\item Smsy - median estimates for the Smsy parameter
\item Umsy - median estimates for the Umsy parameter
\item sigma - median estimates for the sigma parameter
\item rho - median estimates for the rho parameter
\item residuals - median estimates for each epsilon (residual) parameter
\item summary - summary of stan model fit
\item full_posterior - all posterior samples of the above parameters
\item stanfit - the stanfit model object
}
}
\description{
Simple Ricker model estimated with stan
}
\examples{
data(harck)
ricker_stan(data=harck)

}
