% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.R
\name{ricker_TMB}
\alias{ricker_TMB}
\title{Simple Ricker model estimated with TMB}
\usage{
ricker_TMB(
  data,
  silent = FALSE,
  control = TMBcontrol(),
  tmb_map = list(),
  ac = FALSE,
  priors_flag = 1,
  stan_flag = 0,
  sig_p_sd = 1,
  Smax_mean = 230000,
  Smax_sd = 230000
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from \code{\link[=TMBcontrol]{TMBcontrol()}} function, to be passed to nlminb()}

\item{tmb_map}{optional, mapping list indicating if parameters should be estimated of fixed.
Default is all parameters are estimated}

\item{ac}{Logical. Are residuals autocorrelated? Default is FALSE}

\item{priors_flag}{Integer, 1 priors are included in estimation model, 0 priors are not included.
See details for priors documentation. See details for priors documentation.}

\item{stan_flag}{Integer, flag indicating wether or not TMB code will be used with TMBstan - Jacobian
adjustment implemented. Default is 0, jacobian adjustment not included.}

\item{sig_p_sd}{sd for half normal prior on sigma parameter. default is 1.}

\item{Smax_mean}{mean for lognormal prior on Smax, default is 230000. Note that transformation to log space is done internally with log_prior_params()}

\item{Smax_sd}{sd for lognormal prior on Smax, default is 230000. Note that transformation to log space is done internally with log_prior_params()}
}
\value{
a list containing several model outputs:
\itemize{
\item logalpha - MLE estimates for the alpha parameter vector
\item beta - MLE estimates for the beta parameter
\item sigma - MLE estimates for the model standard deviation sigma
\item sigma_noar - if AC= TRUE, MLE estimates for the unadjusted standard deviation sigma
\item rho - if AC=TRUE, MLE estimates for autocorrelation coefficient
\item Smsy - Estimates of Smsy based on estimated logalpha and Smax parameters
\item Umsy -Estimates of umsy based on estimated logalpha parameter
\item residuals - observed - predicted logRS estimates
\item AICc - AICc values, given by 2\emph{nll + 2}npar +(2\emph{npar}(npar+1)/(nrow(data)-npar-1)), excluding prior components
\item BIC - BIC values, excluding prior components
\item model - opt object, generated by the \code{stats::nlminb()}
\item tmb_data - data provided to TMB,
\item tmb_params - parameter intial guesses,
\item tmb_map - mapping indicating which parameters should be fixed or estimated,
\item tmb_obj - obj object, generated by the \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}
\item gradients - final convergence gradient
\item bad_eig - eigen value
\item call - original function call
\item sd_report - MLE estimates and sdt Error estimates for main parameter estimates
\item class - name of cpp model used
}
}
\description{
Simple Ricker model estimated with TMB
}
\details{
Priors: Weakly informative priors are included for the main parameters of the model:
logalpha ~ N(1.5,2.5)
logsmax ~ logN(Smax_mean,Smax_sd)
sigobs ~ halfN(0,sig_p_sd)
}
\examples{

data(harck)
rickerTMB(data=harck)

}
