% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBstan_functions.R
\name{ricker_TMBstan}
\alias{ricker_TMBstan}
\title{Simple Ricker model estimated with TMBstan}
\usage{
ricker_TMBstan(
  data,
  silent = FALSE,
  control = TMBcontrol(),
  tmb_map = list(),
  AC = FALSE,
  priors_flag = 1,
  stan_flag = 0,
  sig_p_sd = 1,
  chains = 6,
  iter = 5000,
  warmup = floor(iter/2)
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from TMBcontrol() function, to be passed to nlminb()}

\item{tmb_map}{optional, mapping list indicating if parameters should be estimated of fixed.
Default is all parameters are estimated}

\item{AC}{Logical. Are residuals autocorrelated? Default is FALSE}

\item{priors_flag}{Integer, 1 priors are included in estimation model, 0 priors are not included.
See details for priors documentation. See details for priors documentation.}

\item{stan_flag}{Integer, flag indicating wether or not TMB code will be used with TMBstan - Jacobian
adjustment implemented. Default is 0, jacobian adjustment not included.}

\item{sig_p_sd}{sd for half normal prior on sigma parameter. default is 1.}

\item{chains}{number of MCMC chains}

\item{iter}{number of MCMC iterations per chain}
}
\value{
a list containing several model outputs - posterior distributions:
\itemize{
\item alpha - posterior distribution for the alpha parameter vector
\item beta - posterior distribution for the beta parameter
\item sigobs - posterior distribution for the observation error sigma
\item Smax - posterior distribution for the Smax parameter vector
\item umsy - posterior distribution for the umsy parameter
\item Smsy - posterior distribution for the Smsy
\item AICc - AICc values, given by 2\emph{nll + 2}npar +(2\emph{npar}(npar+1)/(nrow(data)-npar-1)), excluding prior components
\item BIC - BIC values, excluding prior components
}
}
\description{
Simple Ricker model estimated with TMBstan
}
\details{
Priors: Weakly informative priors are included for the main parameterst of the model:
alpha ~ gamma(3,1)
logbeta ~ N(-12,3)
sigobs ~ gamma(2,1/3)
}
\examples{
data(harck)
rickerTMB(data=harck)

}
