% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.R
\name{rickerTMB}
\alias{rickerTMB}
\title{Simple Ricker model estimated with TMB}
\usage{
rickerTMB(data, silent = FALSE, control = TMBcontrol())
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from TMBcontrol() function, to be passed to nlminb()}
}
\description{
Simple Ricker model estimated with TMB
}
