% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{rw_sr_plot}
\alias{rw_sr_plot}
\title{rw_sr_plot function}
\usage{
rw_sr_plot(
  data,
  mod,
  title = NULL,
  make.pdf = FALSE,
  fig.pars = c(6, 8),
  sr.only = FALSE,
  freq.pred = 3
)
}
\arguments{
\item{data}{stock-recruitment dataset. With columns R (recruits) and S (spawners) and by (brood year)}

\item{mod}{a fitted Stan or TMB static model}

\item{title}{optional title for plot}

\item{make.pdf}{TRUE or FALSE, indicating whether to create a pdf or not, default is FALSE}

\item{sr.only}{TRUE or FALSE, indicating whether to produce just the S-R curve rather than two plots with the spawner-recruit curve and the trajectory of the time-varying parameter (log(alpha) or Smax), default is FALSE}

\item{freq.pred}{frequency of the predicted S-R curve through time. Default = 3, plot a unique curve for every 3 years.}
}
\value{
returns the specified plot(s)
}
\description{
This function plots a time-varying spawner-recruit curves relative to specified spawner-recruit data based on a TMB or Stan 'random walk' model fit from samEst.
}
\examples{
data(harck)
fit=ricker_rw_TMB(data=harck,tv.par='a')
rw_sr_plot(data=harck,mod=fit,title='Harrison Chinook')
}
