% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_functions.R
\name{sGenCalc}
\alias{sGenCalc}
\title{compute Sgen estimate}
\usage{
sGenCalc(a, b, Smsy)
}
\arguments{
\item{a}{alpha parameter in Ricker function: R=S\emph{exp(a-b}S)}

\item{b}{beta parameter in Ricker function: R=S\emph{exp(a-b}S)}

\item{Smsy}{estimate of Smsy based on the alpha and beta parameters above.}
}
\value{
Sgen estimate
}
\description{
compute Sgen estimate
}
\examples{
m1<- ricker_TMB(data=harck)
sgen<-sGenCalc(m1$alpha,m1$beta,m1$Smsy)


}
