% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_functions.R
\name{sgenCalcDirect}
\alias{sgenCalcDirect}
\title{compute Sgen estimate based on the lambertW function similarly to
the approach in Scheuerell 2016. An explicit solution for
calculating optimum spawning stock size from Ricker’s stock recruitment model}
\usage{
sgenCalcDirect(loga, b)
}
\arguments{
\item{loga}{alpha parameter in Ricker function: R=S\emph{exp(loga-b}S)}

\item{b}{beta parameter in Ricker function: R=S\emph{exp(loga-b}S)}
}
\value{
DirectSgen estimate
}
\description{
compute Sgen estimate based on the lambertW function similarly to
the approach in Scheuerell 2016. An explicit solution for
calculating optimum spawning stock size from Ricker’s stock recruitment model
}
