% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_functions.R
\name{Sgencompute}
\alias{Sgencompute}
\title{compute Sgen likelihood function}
\usage{
Sgencompute(S, a, b, Smsy)
}
\arguments{
\item{S}{Spawner numbers set to the interval between 0 and Smsy in sGenSolver.}

\item{a}{alpha parameter in Ricker function: R=S\emph{exp(a-b}S)}

\item{b}{beta parameter in Ricker function: R=S\emph{exp(a-b}S)}

\item{Smsy}{estimate of Smsy based on the alpha and beta parameters above.}
}
\value{
Sgen likelihood
}
\description{
compute Sgen likelihood function
}
