% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{sr_format}
\alias{sr_format}
\title{Function to format the names for a stock-recruitment dataset to avoid compatibility problems with other samEst functions}
\usage{
sr_format(S, R, by)
}
\arguments{
\item{S}{a vector of spawners}

\item{R}{a vector of years for each cohort}
}
\value{
mean of exponentials
}
\description{
Function to format the names for a stock-recruitment dataset to avoid compatibility problems with other samEst functions
}
