% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{sr_mod2}
\alias{sr_mod2}
\title{sr_mod2 function}
\usage{
sr_mod2(
  type = c("static", "rw", "hmm"),
  ac = FALSE,
  par = c("n", "a", "b", "both"),
  lfo = FALSE,
  modelcode = FALSE
)
}
\arguments{
\item{type}{Specify whether to generate a 'static' S-R model, where parameters are time-invariant,
a time-varying 'rw' model, or a regime shift hidden Markov model 'hmm'}

\item{ac}{TRUE or FALSE statement to include autocorrelated residuals. Only compatible with static model}

\item{par}{For time-varying or regime S-R models, what parameter should vary? Either productivity (intercept, a), capacity (slope, b) or both parameters}

\item{lfo}{TRUE or FALSE statement that dictates whether model is being used for out-of-sample log-likelihood estimation}

\item{modelcode}{Logical indicating whether to output model_code or a stan_model object (FALSE, the default)}
}
\value{
returns the compiled rstan code for a given S-R model
}
\description{
Same as 'sr_mod' but does not include the lambert function for reference points (U_msy and)
}
\examples{
m2=sr_mod(type='static',ac = TRUE,par='n',lfo=T)
}
