% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sr_plot}
\alias{sr_plot}
\title{sr_plot function}
\usage{
sr_plot(
  type = c("static", "rw", "hmm"),
  df,
  form = c("stan", "tmb"),
  mod,
  pdf = 0
)
}
\arguments{
\item{type}{Specify whether to generate a 'static' S-R model, where parameters are time-invariant,
a time-varying 'tv' model, or a regime shift model 'regime'}

\item{df}{stock-recruitment dataset. With columns R (recruits) and S (spawners) and by (brood year)}

\item{form}{Either fit with Stan ('stan') or TMB ('tmb')}

\item{mod}{a fitted Stan or TMB model}

\item{pdf}{whether to create a pdf or not}
}
\value{
returns the compiled rstan code for a given S-R model
}
\description{
This function generates a stock-recruitment (S-R) model for rstan based on model inputs.
}
\examples{
sr_plot(type='static',df=df,form='stan',par='n',loglik=T)
}
