% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_stan_functions.R
\name{stan_aic}
\alias{stan_aic}
\title{AIC for stan models}
\usage{
stan_aic(x, form = c("aic", "bic"), type = c("full", "d90", "d80"), k)
}
\arguments{
\item{x}{a list of pointwise out-of-sample loglikelihoods from different models, where each row represents model likelihood predictions}

\item{form}{option for either AIC pr BIC}

\item{type}{option on what observations to include: full = entire out of sample years, d90 = excluding the 10\% of years with the lowest likelihood among all models, d80 = excluding the 20\% of years with the lowest likelihood among all models.}
}
\value{
returns an AIC/BIC score for each model
}
\description{
This function estimates either AIC/BIC scores based on pointwise log likelihood - taking the average across the posterior.
}
