% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_stan_functions.R
\name{stan_compile}
\alias{stan_compile}
\title{stan compile - compiles stan versions of static, autocorrelated, random walk, and regime shift models}
\usage{
stan_compile(type = c("all"), par = c("prod", "cap", "both"))
}
\arguments{
\item{type}{Options for:
\itemize{
\item all (default) - compiles 8 model sets
\item static - compiles the classic (static) Ricker model
\item ac - compiles the static Ricker model with 1-y residual autocorrelation
\item rw - random walk (time-varying) model, must also specify the parameter ('par') - see below
\item hmm - hidden markov (regime shift) model, must also specify the parameter ('par') - see below
}}

\item{par}{Options for:
\itemize{
\item prod - sets stock productivity to vary through time or regime
\item cap - sets stock capacity (Smax) to vary through time or regime
\item both - sets both stock productivity and capacity to vary through time or regime (generally not recommended)
}}
}
\value{
Nothing. It will compile stan code so these models can be sampled, to be used before 'stan_fit' (see help)
}
\description{
stan compile - compiles stan versions of static, autocorrelated, random walk, and regime shift models
}
\examples{
stan_compile()
stan_compile(type='rw',par='prod')
}
