% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_stan_functions.R
\name{stan_format}
\alias{stan_format}
\title{stan_format - formats data for stan models}
\usage{
stan_format(s, r, t, K = 2, smax_priors = NA)
}
\arguments{
\item{s}{Time-series of spawner abundance}

\item{r}{Time-series of recruitment}

\item{t}{Time-series of brood years}

\item{K}{number of regimes for the hidden markov model - dfeault is 2 (high and low productivity or capacity)}

\item{smax_priors}{Custom priors for Smax - these should be two values (e.g. smax_priors=c(1000, 1000)) corresponding to the prior mean and prior error term.}
}
\value{
A list with all necessary inputs for the 8 different stan models
}
\description{
stan_format - formats data for stan models
}
\examples{
stan_compile()
dl=stan_format(R=data$recruits,S=data$spawners,t=data$broodyear)
}
