% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_stan_functions.R
\name{stan_lfo_cv}
\alias{stan_lfo_cv}
\title{stan_lfo_cv function}
\usage{
stan_lfo_cv(
  mod,
  type = c("static", "tv", "regime"),
  df,
  L = 10,
  K = 2,
  dirichlet_prior = NULL
)
}
\arguments{
\item{mod}{Compiled stan model, which can be generated with sr_mod function.}

\item{type}{The type of stock-recruitment model that is being fit. Either 'static' for the time-invariant model (model 1: static Ricker),
'tv' for time-varying models (including autocorrelated residual model), or 'regime' for regime shift models}

\item{df}{Dataframe to use}

\item{L}{Minimum dataset to retain, default is to 10 years of data}

\item{K}{Number of potential regime shifts - default 2}

\item{dirichlet_prior}{k_regime x k_regime matrix. Prior for transition probability matrix,
if NULL prior is set to matrix(1,nrow=k_regime,ncol=k_regime)}
}
\value{
returns the pointwise out-of-sample log likelihoods. For 'tv' or 'regime' models returns a vector that uses either 1. the last year's time-varying parameters,
2. an average of the last 3 years of the time-varying parameters, or 5. an average of the last 5 years of the time-varying parameters. For 'regime' models also includes
these estimates that are also probability-weighted (ie. prob regime x regime parameter) in addition to the parameters for the most likely state 1, 3 or 5 years back.
}
\description{
This function implements exact Leave-Future-Out Cross-Validation (LFO-CV) (see https://arxiv.org/abs/1902.06281),
which will iteratively estimate model likelihood in an out-of-sample context using past data as a training set.
}
\examples{
r=stan_refit(sm=mod3,newdata=df,oos=12)
}
