% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_stan_functions.R
\name{stan_lfo_cv2}
\alias{stan_lfo_cv2}
\title{stan_lfo_cv function}
\usage{
stan_lfo_cv2(
  mod,
  type = c("static", "tv", "regime"),
  df,
  L = 10,
  K = 2,
  dirichlet_prior = NULL
)
}
\description{
stan_lfo_cv function
}
