% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_stan_functions.R
\name{stan_refit}
\alias{stan_refit}
\title{stan_refit function}
\usage{
stan_refit(sm, newdata, oos, regime = FALSE, K = NULL)
}
\arguments{
\item{sm}{Compiled stan model, which can be generated with sr_mod function.}

\item{newdata}{new data to feed to refit models.}

\item{oos}{What row of the new data should be treated as the out-of-sample test}

\item{regime}{TRUE or FALSE statement - is this a regime shift model or not}

\item{K}{Number of potential regime shifts - default NULL}
}
\value{
returns the model fit
}
\description{
This function refits a stan model (via rstan) to iteratively estimate the out-of-sample likelihood
}
\examples{
r=stan_refit(sm=mod3,newdata=df,oos=12)
}
