% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_run_functions.R
\name{stancontrol}
\alias{stancontrol}
\title{Sampling control options.}
\usage{
stancontrol(adapt_delta = 0.99, ...)
}
\arguments{
\item{...}{Anything else. See the 'Control parameters' section of
\link[rstan:stanmodel-method-sampling]{rstan::sampling}.}

\item{eval.max}{Maximum number of evaluations of the objective function
allowed.}

\item{iter.max}{Maximum number of iterations allowed.}
}
\description{
Any arguments to pass to \link[rstan:stanmodel-method-sampling]{rstan::sampling}.
}
