% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{static_sr_plot}
\alias{static_sr_plot}
\title{static_sr_plot function}
\usage{
static_sr_plot(
  data,
  mod,
  title = NULL,
  make.pdf = FALSE,
  fig.pars = c(6, 4),
  plot.params = TRUE,
  sr.only = FALSE
)
}
\arguments{
\item{data}{stock-recruitment dataset. With columns R (recruits) and S (spawners) and by (brood year)}

\item{mod}{a fitted Stan or TMB static model}

\item{title}{optional title for plot}

\item{make.pdf}{TRUE or FALSE, indicating whether to create a pdf or not, default is FALSE}

\item{plot.params}{TRUE or FALSE, indicating whether to plot Ricker parameter values onto the plot, default is TRUE}

\item{sr.only}{TRUE or FALSE, indicating whether to produce just the S-R curve rather than two plots with the spawner-recruit curve and the residuals of the fit over time, default is FALSE}
}
\value{
returns the specified plot(s)
}
\description{
This function plots a spawner-recruit curve relative to specified spawner-recruit data, and residuals through time, based on a TMB or Stan static model fit from samEst.
}
\examples{
data(harck)
fit=ricker_TMB(data=harck)
static_sr_plot(data=harck,mod=fit,title='Harrison Chinook',plot.params=TRUE)
}
