% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_TMB_functions.R
\name{tmb_mod_lfo_cv}
\alias{tmb_mod_lfo_cv}
\title{Simple Ricker model estimated with TMB}
\usage{
tmb_mod_lfo_cv(
  data,
  model = c("static", "staticAC", "rw_a", "rw_b", "rw_both", "HMM", "HMM_a", "HMM_b"),
  L = 10,
  siglfo = c("obs", "total")[1],
  dirichlet_prior = NULL,
  priors_flag = 1,
  sig_p_sd = 1,
  priorSmax = c("default", "maxobsS")[1],
  Smax_mean = exp(12),
  Smax_sd = exp(12)
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{model}{string. Which parameters are time-varying? Options are c('static','rw_a','rw_b',
'rw_both', 'HMM', 'HMM_a', 'HMM_b')}

\item{L}{starting point for LFO-CV (minimum value is 10)}

\item{siglfo}{string. Indicating whether full variance should be used for lfo of models with random walks in parameters
"obs" incates that only observation variance is considered for lfo calculations, "total" indicates that sum of
process and observation variances are used. Option valid only for 'logalpha' tv par.}

\item{dirichlet_prior}{Prior for transition probability matrix in HMM models
k_regime x k_regime matrix. If NULL prior is set to matrix(1,nrow=k_regime,ncol=k_regime).}

\item{priors_flag}{Integer, 1 priors are included in estimation model, 0 priors are not included.}

\item{sig_p_sd}{sd for half normal prior on sigma parameter. default is 1.}

\item{priorSmax}{Options are : "default" or "maxobsS", default will rely on fixed values provided by the user,
"maxobsS" will be calculated based on the observed data.}
}
\value{
vector of lfo by year
}
\description{
Simple Ricker model estimated with TMB
}
\examples{
data(harck)

lfostatic<-tmb_mod_lfo_cv(data=harck,model='static', L=round((2/3)*nrow(harck)))
lfoac <- tmb_mod_lfo_cv(data=harck,model='staticAC', L=round((2/3)*nrow(harck)))
lfoalpha <- tmb_mod_lfo_cv(data=harck,model='rw_a', siglfo="obs", L=round((2/3)*nrow(harck)))
lfobeta <- tmb_mod_lfo_cv(data=harck,model='rw_b', siglfo="obs", L=round((2/3)*nrow(harck)))
lfoalphabeta <- tmb_mod_lfo_cv(data=harck,model='rw_both', siglfo="obs", L=round((2/3)*nrow(harck)))
lfohmma <- tmb_mod_lfo_cv(data=harck,model='HMM_a', L=round((2/3)*nrow(harck)))
lfohmmb <- tmb_mod_lfo_cv(data=harck,model='HMM_b', L=round((2/3)*nrow(harck)))
lfohmm <- tmb_mod_lfo_cv(data=harck,model='HMM', L=round((2/3)*nrow(harck)))


}
