% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP_functions.R
\name{umsySolver}
\alias{umsySolver}
\title{compute Umsy estimate based on Scheuerell 2016. An explicit solution for
calculating optimum spawning stock size from Ricker’s stock recruitment model}
\usage{
umsySolver(a)
}
\arguments{
\item{a}{alpha parameter in Ricker function: R=S\emph{exp(a-b}S)}

\item{b}{beta parameter in Ricker function: R=S\emph{exp(a-b}S)}

\item{Smsy}{estimate of Smsy based on the alpha and beta parameters above.}
}
\value{
Smsy estimate
}
\description{
compute Umsy estimate based on Scheuerell 2016. An explicit solution for
calculating optimum spawning stock size from Ricker’s stock recruitment model
}
