% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{post_check}
\alias{post_check}
\title{Posterior predictive check plot}
\usage{
post_check(fit, data)
}
\arguments{
\item{fit}{A model fit in rstan from e.g. ricker_stan, ricker_rw_stan, ricker_hmm_stan}

\item{data}{the data that were used for the model in 'fit'}
}
\description{
Posterior predictive check plot
}
