% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMB_functions.R
\name{ricker_TMB}
\alias{ricker_TMB}
\title{Simple Ricker model estimated with TMB}
\usage{
ricker_TMB(
  data,
  silent = FALSE,
  control = TMBcontrol(),
  tmb_map = list(),
  AC = FALSE,
  priors = 1
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{silent}{Logical Silent or optimization details? default is FALSE}

\item{control}{output from TMBcontrol() function, to be passed to nlminb()}

\item{tmb_map}{optional, mapping list indicating if parameters should be estimated of fixed.
Default is all parameters are estimated}

\item{AC}{Logical. Are residuals autocorrelated? Default is FALSE}

\item{priors}{Integer, 1 priors are included in estimation model, 0 priors are not included.
See details for priors documentation}
}
\value{
a list containing several model outputs:
\itemize{
\item alpha - MLE estimates for the alpha parameter vector
\item beta - MLE estimates for the beta parameter
\item sigobs - MLE estimates for the observation error sigma
\item AICc - AICc values, given by 2\emph{nll + 2}npar +(2\emph{npar}(npar+1)/(nrow(data)-npar-1)), excluding prior components
\item BIC - BIC values, excluding prior components
\item model - opt object, generated by the \code{stats::nlminb()}
\item tmb_data - data provided to TMB,
\item tmb_params - parameter intial guesses,
\item tmb_map - mapping indicating which parameters should be fixed or estimated,
\item tmb_obj - obj object, generated by the \code{TMB::MakeADFun()}
\item gradients - final convergence gradient
\item bad_eig - eigen value
\item call - original function call
\item sd_report - MLE estimates and sdt Error estimates for main parameter estimates
\item class - name of cpp model
}
}
\description{
Simple Ricker model estimated with TMB
}
\details{
Priors: Weakly informative priors are included for the main parameterst of the model:
alpha ~ N(0,2.5)
logbeta ~ N(-12,3)
sigobs ~ gamma(2,1/3)
}
\examples{
data(harck)
rickerTMB(data=harck)

}
