% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_functions.R
\name{sr_mod}
\alias{sr_mod}
\title{sr_mod function}
\usage{
sr_mod(
  type = c("static", "rw", "hmm"),
  ac = FALSE,
  par = c("n", "a", "b", "both"),
  lfo = FALSE,
  modelcode = FALSE
)
}
\arguments{
\item{type}{Specify whether to generate a 'static' S-R model, where parameters are time-invariant,
a time-varying 'tv' model, or a regime shift model 'regime'}

\item{ac}{TRUE or FALSE statement to include autocorrelated residuals. Only compatible with static model}

\item{par}{For time-varying or regime S-R models, what parameter should vary? Either productivity (intercept, a), capacity (slope, b) or both parameters}

\item{lfo}{TRUE or FALSE statement that dictates whether model is being used for out-of-sample log-likelihood estimation}

\item{modelcode}{Logical indicating whether to output model_code or a stan_model object (FALSE, the default)}

\item{caphigh}{TRUE or FALSE statement to indicate whether you are specifying that high stock capacity or low capacity coincides with a high productivity regime. Only compatible with regime (hmm) model where both alpha and beta are varying (both).}
}
\value{
returns the compiled rstan code for a given S-R model
}
\description{
This function generates a stock-recruitment (S-R) model for rstan based on user inputs.
}
\examples{
m2=sr_mod(type='static',ac = TRUE,par='n',loglik=T)
}
