% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{static_sr_plot}
\alias{static_sr_plot}
\title{static.sr.plot function}
\usage{
static_sr_plot(
  data,
  mod,
  title = NULL,
  make.pdf = FALSE,
  fig.pars = c(6, 4),
  plot.params = FALSE,
  sr.only = FALSE
)
}
\arguments{
\item{mod}{a fitted Stan or TMB static model}

\item{make.pdf}{TRUE or FALSE, indicating whether to create a pdf or not}

\item{plot.params}{TRUE or FALSE, indicating whether to plot Ricker parameter values onto the plot}

\item{df}{stock-recruitment dataset. With columns R (recruits) and S (spawners) and by (brood year)}

\item{resids}{TRUE or FALSE, indicating whether to produce two plots with the spawner-recruit curve and the residuals of the fit over time}
}
\value{
returns the compiled rstan code for a given S-R model
}
\description{
This function generates a stock-recruitment (S-R) model for rstan based on model inputs.
}
\examples{
sr_plot(type='static',df=df,form='stan',df=df,mod=f1,pdf=FALSE)
}
