% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{static_sr_plot}
\alias{static_sr_plot}
\title{static.sr.plot function}
\usage{
static_sr_plot(
  df,
  mod,
  title = NULL,
  make.pdf = FALSE,
  fig.pars = c(6, 4),
  plot.params = FALSE,
  resids = FALSE,
  hpd = NULL
)
}
\arguments{
\item{df}{stock-recruitment dataset. With columns R (recruits) and S (spawners) and by (brood year)}

\item{mod}{a fitted Stan or TMB model}

\item{form}{Either fit with Stan ('stan') or TMB ('tmb')}

\item{pdf}{whether to create a pdf or not}
}
\value{
returns the compiled rstan code for a given S-R model
}
\description{
This function generates a stock-recruitment (S-R) model for rstan based on model inputs.
}
\examples{
sr_plot(type='static',df=df,form='stan',df=df,mod=f1,pdf=FALSE)
}
