% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_TMB_functions.R
\name{tmb_mod_lfo_cv}
\alias{tmb_mod_lfo_cv}
\title{Simple Ricker model estimated with TMB}
\usage{
tmb_mod_lfo_cv(data, tv.par = c("static", "alpha", "beta", "both"), L = 10)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{tv.par}{string. Which parameters are time-varying? Options are c('static','alpha','beta','both')}

\item{L}{starting point for LFO-CV (min. 10)}
}
\value{
vector of lfo by year
}
\description{
Simple Ricker model estimated with TMB
}
\examples{
data(harck)
rickerTMB(data=harck)

}
