% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfo_TMB_functions.R
\name{tmb_mod_lfo_cv}
\alias{tmb_mod_lfo_cv}
\title{Simple Ricker model estimated with TMB}
\usage{
tmb_mod_lfo_cv(
  data,
  model = c("static", "staticAC", "rw_a", "rw_b", "rw_both", "HMM", "HMM_a", "HMM_b"),
  L = 10,
  siglfo = c("obs", "total"),
  dirichlet_prior = NULL
)
}
\arguments{
\item{data}{A list or data frame containing Spawners (S) and log(Recruits/Spawners) (logRS) time series.}

\item{model}{string. Which parameters are time-varying? Options are c('static','rw_a','rw_b',
'rw_both', 'HMM', 'HMM_a', 'HMM_b')}

\item{L}{starting point for LFO-CV (minimum value is 10)}

\item{siglfo}{string. Incating whether full variance should be used for lfo of models with random walks in parameters
"obs" incates that only observation variance is considered for lfo calculations, "total" indicates that sum of
process and observation variances are used. Option valud only for 'alpha' tv par.}

\item{dirichlet_prior}{Prior for transition probability matrix in HMM models
k_regime x k_regime matrix. If NULL prior is set to matrix(1,nrow=k_regime,ncol=k_regime).}
}
\value{
vector of lfo by year
}
\description{
Simple Ricker model estimated with TMB
}
\examples{
data(harck)
tmb_mod_lfo_cv(data=harck, model=c('static'))

}
